rebrand scraping as individual scraping and move settings read and write into helpers
